Move endpoint implementations to separate file
